Replace misleading usage of `toHaveProp` in docs
The example usage of `toHaveProp` in the README shows it being used to make assertions about a component's `key` prop. As far as I can tell, you actually can't make assertions about `key` using `toHaveProp` - React treats `key` specially and doesn't expose it as a regular prop. 

I got tripped up by this subtlety today, and was extra confused as to why it wasn't working since the example of `toHaveProp` in the readme was using `key`. Hopefully, this small change will save someone else a few mins down the road.

BTW, if there is interest I'd be happy to take a run at adding a `toHaveKey` matcher in a separate PR.